Initial commit
This is my developer's portfolio in looking for my first job in big
IT-company I've been dreaming for long time.
There I am going to use most popular frontend stack:
- React
- React router
- Typescript
- Redux
- for UI - Material UI library
- for convenient support In-JS styling - TailwindCSS
- for advanced and React-infrustructure-integrated animation managing -
FramerMotion
Probably there are going to be some else auxilary tools, but considering
that it's relatively not very big server-part-less solution, that's all
we need to have.